Улучшена обработка ошибок

+++ presentation/controllers/deleteCommentButtonController.js
@@ -1,10 +1,16 @@
 import commentsService from "/application/services/commentsService.js";
 import repliesController from "./repliesController.js";
+import authChecker from "/application/authChecker.js";
 
 const deleteCommentButtonController = async (comment, postId = null, rootComment = null) => {
     const deleteButton = comment.querySelector(".delete");
 
     deleteButton.addEventListener("click", async () => {
+        if (!authChecker()) {
+            alert("Необходимо войти в аккаунт");
+            return;
+        }
+
         try {
             await commentsService.deleteComment(comment.id);
             if (rootComment) {
@@ -15,8 +21,7 @@ const deleteCommentButtonController = async (comment, postId = null, rootComment
             }
         }
         catch (error) {
-            alert("Не удалось удалить комментарий");
-            console.error(error);
+            alert(error.message);
         }
     });
 }
